--- packages/components/src/Table-next/index.ts
@@ -1,1 +1,0 @@
-export * from './Table';
