+++ packages/components/src/Table/TableCell.tsx
@@ -1,65 +1,41 @@
 import React, { useRef } from 'react';
-import { useCheckbox } from '@react-aria/checkbox';
+import { useTableCell } from '@react-aria/table';
 import { useFocusRing } from '@react-aria/focus';
-import { useTableCell, useTableSelectionCheckbox } from '@react-aria/table';
 import { mergeProps } from '@react-aria/utils';
-import { TableState } from '@react-stately/table';
-import { useToggleState } from '@react-stately/toggle';
-import { Node } from '@react-types/shared';
+import { GridNode } from '@react-types/grid';
 
-import { CSSObject, useStateProps } from '@marigold/system';
+import { Box, useStateProps } from '@marigold/system';
 
-import { Box } from '../Box';
+import { useTableContext } from './Context';
 
-// Props
-// ----------------------------
 export interface TableCellProps {
-  item: Node<object>;
-  state: TableState<object>;
-  /**
-   * Wheter it is a cell with a checkbox or a regular data cell
-   */
-  isSelectionCell?: boolean;
-  css?: CSSObject;
+  cell: GridNode<object>;
 }
 
-// TableCell Component
-// ----------------------------
-export const TableCell = ({
-  item: cell,
-  state,
-  isSelectionCell,
-  css,
-}: TableCellProps) => {
-  const cellRef = useRef(null);
-  const { gridCellProps } = useTableCell({ node: cell }, state, cellRef);
-
-  const { checkboxProps } = useTableSelectionCheckbox(
-    { key: cell.parentKey! },
-    state
-  );
-  const inputRef = useRef(null);
-  const { inputProps } = useCheckbox(
-    checkboxProps,
-    useToggleState(checkboxProps),
-    inputRef
+export const TableCell = ({ cell }: TableCellProps) => {
+  const ref = useRef(null);
+  const { state, styles } = useTableContext();
+  const disabled = state.disabledKeys.has(cell.parentKey!);
+  const { gridCellProps } = useTableCell(
+    {
+      node: cell,
+    },
+    state,
+    ref
   );
 
   const { focusProps, isFocusVisible } = useFocusRing();
-  const stateProps = useStateProps({ focus: isFocusVisible });
+  const stateProps = useStateProps({ disabled, focusVisible: isFocusVisible });
 
   return (
     <Box
       as="td"
-      ref={cellRef}
-      __baseCSS={{
-        textAlign: isSelectionCell ? 'center' : 'left',
-      }}
-      css={css}
+      ref={ref}
+      css={styles.cell}
       {...mergeProps(gridCellProps, focusProps)}
       {...stateProps}
     >
-      {isSelectionCell ? <input {...inputProps} /> : <>{cell.rendered}</>}
+      {cell.rendered}
     </Box>
   );
 };
