--- packages/components/src/Table/TableRowGroup.tsx
@@ -1,19 +1,0 @@
-import React, { ReactNode } from 'react';
-import { useTableRowGroup } from '@react-aria/table';
-
-// Props
-// ----------------------------
-export interface TableRowGroupProps {
-  children?: ReactNode;
-  as: 'thead' | 'tbody' | 'tfoot';
-}
-
-// Component
-// ----------------------------
-export const TableRowGroup = ({
-  as: Element,
-  children,
-}: TableRowGroupProps) => {
-  const { rowGroupProps } = useTableRowGroup();
-  return <Element {...rowGroupProps}>{children}</Element>;
-};
